converted js file to tsx and made ipdates
Converted Javascript files and code to Typescript.
fixed all the new issues that came up with a updated everything so it works now.
updates /qr so it looks better on windows but still need to fix and update it.
