--- app/qr/components/CategorySummary.js
@@ -1,22 +1,0 @@
-import React from 'react';
-
-const CategorySummary = () => {
-  return (
-    <div className="category-summary">
-      <ul>
-        <li>Starters <span>15</span></li>
-        <li>Main Course <span>12</span></li>
-        <li>Beverages <span>06</span></li>
-        <li>Desserts <span>09</span></li>
-        <li>Alcohol <span>22</span></li>
-        <li></li>
-        <li></li>
-        <li></li>
-        <li></li>
-        <li></li>
-      </ul>
-    </div>
-  );
-};
-
-export default CategorySummary;
