--- app/qr/components/DishCard.js
@@ -1,23 +1,0 @@
-import React from 'react';
-import Image from 'next/image'
-
-const DishCard = ({ name, price, rating, description, image }) => {
-  return (
-    <div className="dish-card">
-      <div className="dish-info">
-        <h4>{name}</h4>
-        <hr></hr>
-        <p style={{display:'flex', justifyContent: 'space-between', marginTop: '4%', fontWeight: 'bold'}}>₹{price} <span style={{position: 'absolute', left: '50%', fontWeight:'normal'}}> ⭐{rating}</span></p>
-        <p>{description}</p>
-        
-        
-      </div>
-      <div className="dish-image">
-        <Image src={image} alt={name} />
-        <button>Add</button>
-      </div>
-    </div>
-  );
-};
-
-export default DishCard;
