+++ package.json
@@ -18,8 +18,9 @@
     "clsx": "^2.1.1",
     "lucide-react": "^0.453.0",
     "next": "15.0.1",
-    "react": "19.0.0-rc-69d4b800-20241021",
-    "react-dom": "19.0.0-rc-69d4b800-20241021",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-feather": "^2.0.10",
     "tailwind-merge": "^2.5.4",
     "tailwindcss-animate": "^1.0.7",
     "twc": "file:"
